Add files via upload
here is the full game that we can be played now and the target score is 5 and repeat the game. and player one which is red can play by UP and DOWN and the purple player by q and a for up and down. 

The next I will do is a client-server so we could play together.